Label generated components with feature name
This follows up recent work to label the components shown in
routes so that route changes obviously succeed without a need
to look in the browser JS console. Components will be re-generated
with the labelling in place.

The .clang-format already present in the project did not create
results that maximally match the previously generated code, so
it is also updated here.